Fix issue of Chat and Get Answer

+++ src/api/answer/delete-answer.js
@@ -1,3 +1,4 @@
+const { ObjectID, ObjectId } = require("mongodb");
 const enums = require("../../../json/enums.json");
 const messages = require("../../../json/messages.json");
 
@@ -22,28 +23,71 @@ module.exports = exports = {
         .status(enums.HTTP_CODES.BAD_REQUEST)
         .json(utils.createResponseObject(data4createResponseObject));
     }
-
     try {
-      const deleteAnswer = await global.models.GLOBAL.ANSWER.findOneAndRemove({
-        _id: answerId,
-        answerBy: user._id,
+      const answerExists = await global.models.GLOBAL.ANSWER.findById(answerId);
+      let findQuestion = await global.models.GLOBAL.QUESTION.findOne({
+        _id: answerExists.question,
       });
-      if (deleteAnswer) {
-        const data4createResponseObject = {
-          req: req,
-          result: 0,
-          message: messages.ITEM_DELETED,
-          payload: {},
-          logPayload: false,
-        };
-        res
-          .status(enums.HTTP_CODES.OK)
-          .json(utils.createResponseObject(data4createResponseObject));
+      if (findQuestion) {
+        const id = findQuestion._id;
+        const answerBy = user._id;
+        const questionBy = findQuestion.createdBy;
+
+        let participateIds = [];
+        participateIds.push(answerBy);
+        participateIds.push(id);
+        participateIds.push(questionBy);
+
+        let answerRoom =
+          await global.models.GLOBAL.ANSWER_ROOM.findOneAndUpdate(
+            {
+              participateIds: {
+                $size: participateIds.length,
+                $all: [...participateIds],
+              },
+            },
+            {
+              $pull: {
+                answer: { answerId: ObjectId(answerId) },
+              },
+            },
+            { new: true }
+          );
+        const deleteAnswer = await global.models.GLOBAL.ANSWER.findOneAndRemove(
+          {
+            _id: answerId,
+            answerBy: user._id,
+          }
+        );
+
+        if (deleteAnswer) {
+          const data4createResponseObject = {
+            req: req,
+            result: 0,
+            message: messages.ITEM_DELETED,
+            payload: {},
+            logPayload: false,
+          };
+          res
+            .status(enums.HTTP_CODES.OK)
+            .json(utils.createResponseObject(data4createResponseObject));
+        } else {
+          const data4createResponseObject = {
+            req: req,
+            result: -1,
+            message: messages.NOT_ALLOWED,
+            payload: {},
+            logPayload: false,
+          };
+          res
+            .status(enums.HTTP_CODES.OK)
+            .json(utils.createResponseObject(data4createResponseObject));
+        }
       } else {
         const data4createResponseObject = {
           req: req,
           result: -1,
-          message: messages.NOT_ALLOWED,
+          message: "Sorry, Something went wrong to delete answer.",
           payload: {},
           logPayload: false,
         };
